Bunch of job tests, mocking GraphQLClient.request finally working

+++ src/types/job.js
@@ -3,35 +3,53 @@ import {
   destroy as deleteQuery,
   results as resultsQuery,
 } from '../queries'
-import { UpdateError, DeleteError, ResultsError } from '../errors'
+import {
+  UpdateError,
+  DeleteError,
+  ResultsError,
+  ReadOnlyError,
+} from '../errors'
 import Type from './type'
 import JobResult from './jobResult'
 
 export default class Job extends Type {
   parse(data) {
-    this.name = data.name
-    this.enabled = data.enabled
-    this.body = data.body
-    this.endpoint = data.endpoint
-    this.verb = data.verb
-    this.headers = data.headers
-    this.body = data.body
-    this.retryable = data.retryable
-    this.runAt = data.runAt
-    this.runEvery = data.runEvery
+    this.name = data.name || this.name
+    this.enabled = data.enabled || this.enabled
+    this.body = data.body || this.body
+    this.endpoint = data.hasOwnProperty('endpoint')
+      ? data.endpoint
+      : this.endpoint
+    this.verb = data.verb || this.verb
+    this.headers = data.headers || this.headers
+    this.retryable = data.hasOwnProperty('retryable')
+      ? data.retryable
+      : this.retryable
+    this.runAt = data.runAt ? new Date(data.runAt) : this.runAt || null
+    this.runEvery = data.runEvery || this.runEvery
     this.createdAt = data.createdAt
+      ? new Date(data.createdAt)
+      : this.createdAt || null
     this.updatedAt = data.updatedAt
+      ? new Date(data.updatedAt)
+      : this.updatedAt || null
     this.lastRunAt = data.lastRunAt
+      ? new Date(data.lastRunAt)
+      : this.lastRunAt || null
     this.nextRunAt = data.nextRunAt
+      ? new Date(data.nextRunAt)
+      : this.nextRunAt || null
   }
 
   async update(params) {
+    if (this.isDeleted) throw new ReadOnlyError()
+
     try {
-      const data = await this._request(
+      const data = await this.client.request(
         updateQuery,
         Object.assign({ name: this.name }, params)
       )
-      this._reinitialize(data)
+      this.parse(data.job)
       return this
     } catch (error) {
       return new UpdateError(error.message)
@@ -39,8 +57,11 @@ export default class Job extends Type {
   }
 
   async delete() {
+    if (this.isDeleted) throw new ReadOnlyError()
+
     try {
       await this.client.request(deleteQuery, { name: this.name })
+      this.isDeleted = true
       return this
     } catch (error) {
       throw new DeleteError(error.message)
@@ -48,10 +69,13 @@ export default class Job extends Type {
   }
 
   async results() {
+    if (this.isDeleted) throw new ReadOnlyError()
+    
     try {
       const data = await this.client.request(resultsQuery, {
         jobName: this.name,
       })
+
       return data.jobResults.map(
         (result) =>
           new JobResult(result, {
